Extract publishing into its own gradle script plugin
